requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
pydantic==2.5.0
anthropic==0.7.1
apscheduler==3.10.4
pytz==2023.3
python-multipart==0.0.6
aiofiles==23.2.1
